+++ FSIC/classes/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""
+classes
+=======
+FSIC subpackage of class definitions.
+
+Defines:
+ - Schematic classes: Model, Function, Equation
+
+"""
+
+
+from FSIC.classes.schematic import Equation, Function, Model